refactor: move add_pkg_config to arch
