+++ builder/src/profiling.rs
@@ -3,7 +3,7 @@
 
 use crate::arch;
 use crate::module::Module;
-use crate::utils::{adjust_extern_symbols, file_replace, project_root, wait_for_success};
+use crate::utils::{adjust_extern_symbols, project_root, wait_for_success};
 use anyhow::Result;
 use serde::Deserialize;
 use std::ffi::OsStr;
@@ -118,46 +118,11 @@ impl Profiling {
     }
 
     fn add_pkg_config(&self) -> Result<()> {
-        #[cfg(target_os = "windows")]
-        return Ok(());
-
-        let files: [&str; 3] = [
-            "datadog_profiling.pc",
-            "datadog_profiling_with_rpath.pc",
-            "datadog_profiling-static.pc",
-        ];
-
-        //Create directory
-        let pc_dir = Path::new(self.target_pkconfig.as_ref());
-        fs::create_dir_all(pc_dir).expect("Failed to create pkgconfig directory");
-
-        // Create files
-        for file in files.iter() {
-            let file_in = file.to_string() + ".in";
-
-            let mut pc_origin: PathBuf = project_root();
-            pc_origin.push(CRATE_FOLDER);
-            pc_origin.push(file_in);
-
-            let pc_target: PathBuf = [pc_dir.as_os_str(), OsStr::new(file)].iter().collect();
-
-            file_replace(
-                pc_origin.to_str().unwrap(),
-                pc_target.to_str().unwrap(),
-                "@Datadog_VERSION@",
-                &self.version,
-            )?;
-
-            if *file == files[2] {
-                file_replace(
-                    pc_origin.to_str().unwrap(),
-                    pc_target.to_str().unwrap(),
-                    "@Datadog_LIBRARIES@",
-                    arch::NATIVE_LIBS,
-                )?;
-            }
-        }
-        Ok(())
+        arch::add_pkg_config(
+            CRATE_FOLDER,
+            self.target_pkconfig.as_ref(),
+            self.version.as_ref(),
+        )
     }
 }
